requirements.txt
fastapi
uvicorn
openai
langchain
langchain-community
litellm
python-dotenv
fastapi
pydantic
uvicorn